Prevent comparison of double via == in test

+++ test/optional.t.cpp
@@ -717,7 +717,7 @@ CASE( "optional: Provides relational operators" )
 
 CASE( "optional: Provides mixed-type relational operators" )
 {
-    relop<char, int, double>( lest_env );
+    relop<char, int, long>( lest_env );
 }
 
 CASE( "make_optional: Allows to copy-construct optional" )
